Begin to refactor things
